moved applications tracker to the applications folder
