+++ apps/ca-portal/ui/src/components/applications/ApplicationProgressTracker.js
